+++ jgclark.DashboardReact/changelog.md
@@ -14,6 +14,7 @@ Misc TODO:
 - fixed top-alignment spacing for calendar picker
 - changed project dialog to have one month calendar picker so that it fits in the smaller dialog
 - fixed the #home not showing up if it was off on load
+- added incrementallyRefreshSections to load all sections (other than Today) on first load
 
 ## [2.0.0.a22] @jgc
 - completed layout in Project dialog (apart from CalendarPicker)